# Use fast convolution algorithm in `cupy.poly1d.__pow__`