--- a/cupy/lib/_routines_poly.py
+++ b/cupy/lib/_routines_poly.py
@@ -4,6 +4,7 @@ import warnings
 import numpy
 
 import cupy
+import cupyx.scipy.fft
 
 
 def _wraps_polyroutine(func):
@@ -173,7 +174,6 @@ def _polypow(x, n):
         else:
             fft, ifft = cupy.fft.rfft, cupy.fft.irfft
         out_size = (x.size - 1) * n + 1
-        import cupyx.scipy.fft
         size = cupyx.scipy.fft.next_fast_len(out_size)
         fx = fft(x, size)
         fy = cupy.power(fx, n, fx)