- master
```
__pow__             :    CPU:17129.355 us   +/-742.564 (min:15711.574 / max:18453.500) us     GPU-0:17138.483 us   +/-742.922 (min:15719.424 / max:18463.743) us
```

- This PR
```
__pow__             :    CPU:  218.919 us   +/- 5.220 (min:  213.281 / max:  230.438) us     GPU-0:  227.021 us   +/- 4.782 (min:  221.184 / max:  236.544) us
```

- benchmark
```py
import cupy
from cupy import testing
from cupyx.profiler import benchmark


x = testing.shaped_random((5,), xp=cupy, dtype='f', seed=0, scale=0.4)
x = cupy.poly1d(x)
n = 5000
perf = benchmark(cupy.poly1d.__pow__, (x, n), n_repeat=10)

print(perf)
```